testing: move NoopClientCall & NoopServerCall to internal
