+++ core/src/test/java/io/grpc/ContextsTest.java
@@ -29,7 +29,7 @@ import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
 import io.grpc.internal.FakeClock;
-import io.grpc.testing.NoopServerCall;
+import io.grpc.internal.NoopServerCall;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
