+++ core/src/test/java/io/grpc/ServerInterceptorsTest.java
@@ -30,7 +30,7 @@ import static org.mockito.Mockito.verifyZeroInteractions;
 import io.grpc.MethodDescriptor.Marshaller;
 import io.grpc.MethodDescriptor.MethodType;
 import io.grpc.ServerCall.Listener;
-import io.grpc.testing.NoopServerCall;
+import io.grpc.internal.NoopServerCall;
 import java.io.ByteArrayInputStream;
 import java.io.InputStream;
 import java.util.ArrayList;
