+++ stub/src/test/java/io/grpc/stub/ClientCallsTest.java
@@ -37,9 +37,9 @@ import io.grpc.Status;
 import io.grpc.StatusRuntimeException;
 import io.grpc.inprocess.InProcessChannelBuilder;
 import io.grpc.inprocess.InProcessServerBuilder;
+import io.grpc.internal.NoopClientCall;
 import io.grpc.stub.ServerCalls.NoopStreamObserver;
 import io.grpc.stub.ServerCallsTest.IntegerMarshaller;
-import io.grpc.testing.NoopClientCall;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Iterator;
